categories

+++ main-service/src/main/java/ru/practicum/ewm/main/user/service/UserServiceImpl.java
@@ -2,6 +2,8 @@ package ru.practicum.ewm.main.user.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import ru.practicum.ewm.main.exception.NotFoundException;
+import ru.practicum.ewm.main.exception.ValidationException;
 import ru.practicum.ewm.main.user.dto.UserDto;
 import ru.practicum.ewm.main.user.mapper.UserMapper;
 import ru.practicum.ewm.main.user.repository.UserRepository;
@@ -17,17 +19,25 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public UserDto createUser(UserDto userDto) {
-        return userMapper.toDto(userRepository.save(userMapper.toEntity(userDto)));
+        if (!userRepository.existsByEmailIgnoreCase(userDto.getEmail())) {
+            return userMapper.toDto(userRepository.save(userMapper.toEntity(userDto)));
+        } else {
+            throw new ValidationException("Email " + userDto.getEmail() + " alredy in use");
+        }
     }
 
     @Override
     public void deleteUserById(Long userId) {
-        userRepository.deleteById(userId);
+        if (userRepository.existsById(userId)) {
+            userRepository.deleteById(userId);
+        } else {
+            throw new NotFoundException("User with id " + userId + " not found");
+        }
     }
 
     @Override
     public List<UserDto> getUsers(int from, int size) {
-        return List.of();
+        return userMapper.toDtoList(userRepository.findUsers(from, size));
     }
 
     @Override
